Add target to build mods
Requires migration.

This adds the ability to designate the target of particular mods in a build as being either client, server or both (universal). There are no changes to the API yet though, and a call to retrieve all mods for a build will still only return mods with the client or universal target set. Universal is the default option and will be set on all existing builds during migration.